pyproject.toml
[project]
name = "tellmeastory"
version = "0.1.0"
description = "A novel-writing agent that streams prose from a story concept"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
tellmeastory = "tellmeastory.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tellmeastory"]
